pyproject.toml
[project]
name = "forex-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-task-sdk>=1.1.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "supabase>=2.27.0",
    "twelvedata>=1.2.25",
]
